backend/requirements.txt
flake8==4.0.1
Django==3.2.18
djangorestframework==3.14
django-filter==21.1
djoser==2.1.0
isort==5.10.1
python-decouple==3.5
drf-extra-fields==3.2.1
gunicorn==20.1.0
Pillow==9.4.0
reportlab==3.6.12
psycopg2-binary==2.9.5
python-dotenv==0.21.0
django-colorfield==0.8.0
django-cors-headers==3.13.0